+++ packages/playground/CHANGELOG.md
@@ -1,5 +1,12 @@
 # playground
 
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies
+    - @1adybug/react-amap@0.3.1
+
 ## 1.0.4
 
 ### Patch Changes